Flatten SharedPreferencesCustomTileProviderRepository
